requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
asyncmy==0.2.10
asyncodbc==0.1.1
asyncpg==0.30.0
attrs==24.3.0
Brotli==1.1.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
exceptiongroup==1.2.2
fastapi==0.115.6
frozenlist==1.5.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
hyperframe==6.0.1
idna==3.10
iso8601==2.1.0
loguru==0.7.3
msgpack==1.1.0
multidict==6.1.0
nonebot-adapter-onebot==2.4.6
nonebot-plugin-localstore==0.7.3
nonebot-plugin-tortoise-orm==0.1.4
nonebot2==2.4.1
nonestorage==0.1.0
propcache==0.2.1
pycares==4.5.0
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
pygtrie==2.5.0
pyodbc==5.2.0
pypika-tortoise==0.3.2
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.1
starlette==0.41.3
tomli==2.2.1
tortoise-orm==0.23.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==14.1
yarl==1.18.3
